backend/requirements.txt
Flask==3.1.1
google-api-python-client==2.*
google-auth==2.*
google-auth-oauthlib==1.*
dateparser==1.*
python-dotenv==1.*
